requirements.txt
selenium==3.141.0
chromedriver-binary==77.0.3865.40.0
pandas==0.25.3
Flask==1.0.2
gunicorn==19.9.0
google-cloud-storage
webdriver-manager
